+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support FEP-c390 identity proofs.
+- Allow to pin posts to profile.
 
 ### Changed